+++ modulo2/apis-labenusers/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "axios": "^0.27.2"
+  }
+}
